Change to use new input file format

+++ src/main/java/rui/classifier/bayes/BayesReducer.java
@@ -1,8 +1,6 @@
 package rui.classifier.bayes;
 
 import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.io.Text;
 
@@ -13,8 +11,8 @@ import com.google.gson.JsonElement;
 public class BayesReducer extends Reducer<Text, Text, Text, Text> {
     @Override
     public void reduce(Text token, Iterable<Text> records, Context context) throws IOException, InterruptedException {
-        int negativeSum = 0;
-        int positiveSum = 0;
+        long negativeSum = 0;
+        long positiveSum = 0;
 
         for (Text record : records) {
             JsonParser jsonParser = new JsonParser();
@@ -25,6 +23,7 @@ public class BayesReducer extends Reducer<Text, Text, Text, Text> {
                 negativeSum += jsonObject.get("negative").getAsInt();
             }
         }
+        
         JsonObject sumRecord = new JsonObject();
         sumRecord.addProperty("positive", positiveSum);
         sumRecord.addProperty("negative", negativeSum);
